// src/types/Product.ts
export interface Product {
  id?: string;
  name: string;
  category: string;
  code: string;
  price: number;
  discountedPrice: number;
  details: string;
  size: string;
  color: string;
  brand: string;
  tag: string;
  stock: number;
  availability: string;
  thumbnail: string;
  image1: string;
  image2: string;
  image3: string;
  minimumOrderQuantity: number;
}